add location

+++ data-quiz-list/src/main/java/ru/dansh1nv/quiz/data/datasource/quizPlease/QuizPleaseRemoteDataSource.kt
@@ -8,7 +8,7 @@ class QuizPleaseRemoteDataSource(
     private val api: QuizPleaseApi,
 ) {
     suspend fun getQuizList(
-        cityId: Int,
+        cityId: Long,
         pageNumber: Int,
         pageSize: Int,
     ): Flow<List<QuizPleaseDTO>> = api.getQuizzes(
