+++ domain-quiz-list/src/main/java/ru/dansh1nv/quiz_list_domain/repository/IQuizPleaseRepository.kt
@@ -6,7 +6,7 @@ import ru.dansh1nv.quiz_list_domain.models.QuizPlease
 interface IQuizPleaseRepository {
 
     suspend fun getQuizList(
-        cityId: Int,
+        cityId: Long,
         pageNumber: Int,
         pageSize: Int,
     ): Flow<List<QuizPlease>>
