+++ quizapi/src/main/java/ru/dansh1nv/quizapi/api/QuizPleaseApi.kt
@@ -17,7 +17,7 @@ class QuizPleaseApi(
 ) {
 
     suspend fun getQuizzes(
-        cityId: Int,
+        cityId: Long,
         pageNumber: Int,
         pageSize: Int,
     ) = flow {
